refactor: move schema validator to introspection
