+++ packages/integration-boilerplate-node/src/introspectionHandler.ts
@@ -6,7 +6,50 @@ import {
   UnhandledIntrospectionError,
 } from "./errors";
 import { IntegrationService } from "./service";
-import { IntrospectionBody, validateIntrospectionBody } from "./types";
+import { IntrospectionBody } from "./types";
+import Ajv from "ajv";
+import addFormats from "ajv-formats";
+import * as draft6MetaSchema from "ajv/dist/refs/json-schema-draft-06.json";
+import * as integrationTemplateSchema_import from "@saasquatch/schema/json/IntegrationConfig.schema.json";
+
+const additionalPropertiesTrue = (input: any): void => {
+  Object.entries(input).forEach(([key, val]) => {
+    if (key === "additionalProperties") {
+      input[key] = true;
+    } else if (typeof val === "object") {
+      additionalPropertiesTrue(val);
+    }
+  });
+};
+
+const integrationTemplateSchema = { ...integrationTemplateSchema_import };
+// @ts-ignore
+integrationTemplateSchema["$id"] = "#/definitions/integrationTemplateSchema";
+
+// allow additionalProperties: true throughout the entire schema. this enables
+// "at least equivalent" ajv validation for the input
+additionalPropertiesTrue(integrationTemplateSchema);
+
+const introspectionBodySchema = {
+  $schema: "http://json-schema.org/draft-06/schema#",
+  type: "object",
+  required: ["tenantAlias", "templateIntegrationConfig"],
+  additionalProperties: true,
+  properties: {
+    tenantAlias: { type: "string" },
+    templateIntegrationConfig: {
+      $ref: "#/definitions/integrationTemplateSchema",
+    },
+  },
+};
+
+const ajv = new Ajv();
+addFormats(ajv);
+ajv.addMetaSchema(draft6MetaSchema);
+ajv.addSchema(integrationTemplateSchema);
+ajv.addSchema(introspectionBodySchema);
+
+const validateIntrospectionBody = ajv.compile(introspectionBodySchema);
 
 export async function introspectionHandler<
   ServiceConfig extends BaseConfig,
