+++ packages/integration-boilerplate-node/src/types.ts
@@ -1,12 +1,5 @@
-import { JSONSchema6 } from "json-schema";
-import Ajv from "ajv";
-import addFormats from "ajv-formats";
-import * as draft6MetaSchema from "ajv/dist/refs/json-schema-draft-06.json";
-import * as integrationTemplateSchema_import from "@saasquatch/schema/json/IntegrationConfig.schema.json";
-
 import { IntegrationConfiguration } from "@saasquatch/schema/types/IntegrationConfig";
-
-const integrationTemplateSchema = { ...integrationTemplateSchema_import };
+import { JSONSchema6 } from "json-schema";
 
 export type TenantAlias = string;
 export interface Webhook {
@@ -17,36 +10,6 @@ export interface Webhook {
   created: number;
   [key: string]: any;
 }
-
-const additionalPropertiesTrue = (input: any): void => {
-  Object.entries(input).forEach(([key, val]) => {
-    if (key === "additionalProperties") {
-      input[key] = true;
-    } else if (typeof val === "object") {
-      additionalPropertiesTrue(val);
-    }
-  });
-};
-
-// @ts-ignore
-integrationTemplateSchema["$id"] = "#/definitions/integrationTemplateSchema";
-
-// allow additionalProperties: true throughout the entire schema. this enables
-// "at least equivalent" ajv validation for the input
-additionalPropertiesTrue(integrationTemplateSchema);
-
-const introspectionBodySchema = {
-  $schema: "http://json-schema.org/draft-06/schema#",
-  type: "object",
-  required: ["tenantAlias", "templateIntegrationConfig"],
-  additionalProperties: true,
-  properties: {
-    tenantAlias: { type: "string" },
-    templateIntegrationConfig: {
-      $ref: "#/definitions/integrationTemplateSchema",
-    },
-  },
-};
 
 export type IntrospectionBody<IntegrationConfig> = {
   /**
@@ -69,14 +32,6 @@ export type IntrospectionResponse = {
    */
   templateIntegrationConfig: IntegrationConfiguration;
 };
-
-const ajv = new Ajv();
-addFormats(ajv);
-ajv.addMetaSchema(draft6MetaSchema);
-ajv.addSchema(integrationTemplateSchema);
-ajv.addSchema(introspectionBodySchema);
-
-export const validateIntrospectionBody = ajv.compile(introspectionBodySchema);
 
 export interface FormRequestContext<IntegrationConfig, FormConfig> {
   type: "SUBMIT" | "VALIDATE" | "INTROSPECTION" | "INITIAL_DATA";
